Move TypeScript files back to lib
