+++ test/test-uint8.ts
@@ -2,7 +2,7 @@
 
 import {} from 'mocha';
 import {assert} from 'chai';
-import * as Token from '..';
+import * as Token from '../lib';
 import * as util from './util';
 
 describe("Parse 8-bit unsigned integer (UINT8)", () => {
