+++ test/tsconfig.json
@@ -1,7 +1,4 @@
 {
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "declaration": false
-  }
+  "extends": "../tsconfig.json"
 }
