+++ tsconfig.json
@@ -5,7 +5,7 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "target": "ES2017",
-    "declaration": true
+    "declaration": false
   }
 }
